Refactor MeterView into scaffolding + meter-specific views
Extract meter-specific implementations from MeterView.swift into dedicated
files and centralize shared types in MeterTypes.swift. Introduces shared
MeterDimensions.widthRatio so all meters use consistent full screen width.

Closes #22
